Fix issue with patching staticmethod w/ staticmethod
Summary: Fixes a small issue I noticed while working on D52525922.  When we go to handle a value being patched we don't do the unwrapping of the static/class method.  So this moves that logic into `update_thunk` and moves the `_PyClassLoader_ResolveFunction` unwrapping to only happen in the case where we're not returning a thunk.

Reviewed By: carljm

Differential Revision: D52527349

fbshipit-source-id: fbe49c2ced72c68b1467a607ad20a2987c90c36c