Clean up how phi operands are created in LIR generation
Summary:
Phi operands are not necessarily defined at the time the LIR generator reaches
the phi instruction.  They may be defined by a block later in the control flow
graph that's passing the value via a backwards branch.  The way the generator
currently handles them is it allocates dummy operands with null defining
instructions.  Similarly there's no access to the LIR blocks for all phi
operands, so the generator does a hacky thing where it casts the HIR block and
adds that as the label operand.  Once the CFG is fully processed, these are
cleaned up by the FixPhiNodes() method.

This can be cleaner by allocating the phi instruction with zero operands when
the generator first gets to it.  Then fill those in in the same way as before,
but loading the HIR data off of the HIR instruction rather than bogus values off
of the LIR instruction.

To do this we currently need to make a second BasicBlockBuilder to get access to
the getDefInstr() method.  Currently the block builder will always allocate a
new entry block in its ctor, this diff breaks that out and moves it to
TranslateOneBasicBlock() to keep the control flow graph the same.

This also renames FixPhiNodes() to resolvePhiOperands() to both fix the style
and make it a bit more explicit as to what the method is doing.

Reviewed By: swtaarrs

Differential Revision: D52805870

fbshipit-source-id: 2485181ddae851570ad0eb528d4ae6f7242379f8