Customize and brand Windows installer build
Summary:
this patches our fork of the Python 3.12 Windows installer, making customization and branding it as "Meta Python".

in particular, customizing the ReleaseURI and various registry keys allows Meta Python to coexist side-by-side with other installed Pythons of the same version (like the 3.12.1 from the official upstream installer) without interfering.

it's unfortunately this is a rather invasive patch, but understandably, the upstream installer is not designed for ease of branding and customization. we can explore extending the upstream installer to better support such branding & customization as a follow-up.

Reviewed By: amyreese, lisroach, aleivag

Differential Revision: D52912035

fbshipit-source-id: 6a68f8212cb399259e7869e41b4462557f83c4be