take control of pyc reading/writing
Summary:
To implement the plan outlined in the design doc in the previous diff, we'll need to take fuller control of pyc file reading and writing. We will no longer just be inserting an extra magic number at the start of the pyc, we will now need to add an extra dependency tuple partway through it.

This diff makes no functional changes (yet), it just copies in a few additional methods and functions from importlib that we will need to modify in future diffs. The copied-in code is not modified in this diff (except to add a few ignores needed to keep the typechecker happy.)

Reviewed By: DinoV

Differential Revision: D52824192

fbshipit-source-id: a99fddaeb19aa01d787676d1a562664148326783