run pyfmt on cinderx/cinderx/*.py
Summary:
This diff just runs `pyfmt` on the `.py` files in `cinderx/cinderx/`, which have been accidentally exempt from linting for a while. The next diff re-enables linting on them.

The only manual change is to remove a couple grouping comments in `cinderx/cinderx/__init__.py` which are no longer accurate once the imports are auto-sorted alphabetically.

Reviewed By: swtaarrs

Differential Revision: D52862474

fbshipit-source-id: fb48d154ccfaecb130730d1aab4d3ee1990fe6ee