Expect PyFunctionObject* in _PyJit_IsCompiled()
Summary:
If you look at most of its callsites, they already have a PyFunctionObject* but
have to cast it to PyObject*.  Within the function, the argument is expected to
be a PyFunctionObject*.

is_jit_compiled() stands out here because unlike jit_suppress(), disassemble(),
and print_hir(), it does not error when it gets an argument that's not a
PyFunctionObject.  It just returns false.  This diff keeps this behavior.

Reviewed By: swtaarrs

Differential Revision: D53007959

fbshipit-source-id: 1d012eae2a5a01d4d7f03b5738e1005df93f9bb9