Map HIR values directly to LIR instructions in LIR generation
Summary:
Right now there's still a string layer between the two, where the `output_map`
in `Environ` is keyed on the string name of an HIR register.  It can instead key
on the register value directly.

I believe this is the last bit where we convert HIR values into string form for
the purposes of LIR generation.  There still exist string labels for basic
blocks though.

Reviewed By: jbower-fb

Differential Revision: D53015570

fbshipit-source-id: 615e5330803d1737bd76498a41dcf3434b4f0709