make StrictModule not inherit ModuleType
Summary:
StrictModules don't have an `md_dict`; their globals are hidden away so as to be inaccessible to mutation. But they also inherit `PyModule_Type`, and CPython expects every instance of `PyModule_Type` to have a non-null `md_dict`. Most of the time, this is fine, but there can be cases in shutdown where `PyModule_Dict` is called on a strict module and hits an assert in a debug build due to the null `md_dict`.

This diff resolves the contradiction by making strict modules no longer inherit `PyModule_Type` at all. This means we have to take a different approach in the `_static` module; we can't use the normal extension module initialization machinery to create something that isn't a PyModule, so we make `_static` a regular module initially, and then after the fact (in `cinderx.init()`) we import it and replace it in `sys.modules` with a strict module with the same contents. This also allows us to remove some Cinder-specific hooks we'd added in `moduleobject.c` to support mutating a StrictModule via C API during its initialization.

Reviewed By: DinoV

Differential Revision: D53066715

fbshipit-source-id: c178218b7c3e4d68462a3a8dc77aff0f28c9df5c