ensure no lazy imports in loader module
Summary:
Any lazy imported name that is referenced within the code that implements import can cause a cycle. So we need to ensure that all names in our module-loader code that are referenced in the critical path of import are always eagerly imported.

The simplest and most robust way to do this is just to place all imports in the `loader` module inside a try/except, to force them all to be eagerly imported.

Reviewed By: Kronuz

Differential Revision: D53230376

fbshipit-source-id: d5057977f08f5c57c3a6c0589e705fabbe379ca5