alias cinderx.compiler to compiler
Summary:
We want to migrate from `from compiler.strict.loader import install` to `from cinderx.compiler.strict.loader import install`. Since these imports exist across multiple codebases, we have to make both kinds of imports work side-by-side for a period of time while we migrate. And since some compiler modules hold mutable state, we need to ensure there aren't multiple copies of any of these modules; it must be an alias, not a duplication.

This is actually quite tricky to do in Python; the only way I could find to do it (short of messing with import loaders, which is going to be difficult to make transparent to downstream users of custom import loaders) is to mirror the entire package directory structure with all `.py` files, and then `import *` from the canonical locations in the mirror `.py` files. I wrote a script to generate these shim modules, which I included in this diff.

Underscore-prefixed names are a bit of a hassle, since they are not imported by `import *`. Rather than write code to enumerate all symbols in each module and import them all explicitly, I just handled the few underscore-prefixed cases manually. In one case (`_TMP_VAR_PREFIX`) the value was imported in a non-test module, so I renamed it to remove the initial underscore, since clearly it is not private to the module. In the other cases I just added manual imports in the mirror module.

Reviewed By: swtaarrs

Differential Revision: D53204699

fbshipit-source-id: 557c98c6577f48b478fc575b4dbd8cc637bd7cc2