prevent unhandled Python exception in SP recursive compile
Summary:
When jit-compiling a Static Python function which contains an invoke of another Static Python function, we immediately attempt a recursive compile of the target function. This gains some perf, because it means we can emit a direct native x64 call with no indirection overhead.

The problem with this is that preloading a function to be jit-compiled can raise Python exceptions, but actual JIT compilation cannot raise Python exceptions, so we have no way to handle a Python exception raised by the preload portion of the recursive compile.

Solve this by ensuring that we preload the target function already during preload of the caller, so that our recursive compile attempt never has to do any preloading.

This means that all forms of compilation (not only multi-threaded batch compile, but also single-threaded batch compile, and single-function compile) now use the global `jit_preloaders` map. In the batch cases, it gets populated with preloaders for everything we intend to compile. In the single-function case, it gets populated with preloaders for the function we are compiling and anything it statically invokes (transitively).

This means in turn that `JITContext::compilePreloader` is the only JITContext entrypoint we use; `compileFunc` and `compileCode` become unused conveniences. So I removed them.

At some point it might be a good refactor to make the preloaders-map not global, but rather per-compile-batch and passed down as needed. But I didn't try to do that in this diff.

Reviewed By: alexmalyshev

Differential Revision: D53020018

fbshipit-source-id: 3e5a303c833386b9bdc9b7b5bb0f8177d26a27d7