For fbcode CinderX root target, add missing dependencies on _strictmodule and _static
Summary:
A warning about these being missing was being printed from the `assert_fbcode__cinderx_so.py` test. I think perhaps I missed this because code to emit the error (as a warning) was added later.

Also a fix when deleting non-existent `_static.__file__`. That code was added recently so I guess nobody got far enough to see the error in this context.

Reviewed By: carljm

Differential Revision: D53358957

fbshipit-source-id: f1bcbee4ad460faa47447a9459f99a5346262cf5