refactor get_code/get_data/set_data
Summary:
Now that we control the `get_code` method, we don't have to jump through so many hoops in `get_data` and `set_data` -- in fact we can eliminate our overrides of these methods and just handle the strict/static logic around bytecode caching directly in `get_code` instead.

Responsibility for reading and checking the strict magic number is moved into `classify_strict_pyc`, and responsibility for writing it is moved into `code_to_strict_*_pyc`, so it's more aligned with how the regular magic number is handled. Responsibility for adding the strict tag to the pyc filename is moved into `get_code`, and into `strict_compile` when using `compileall`.

Reviewed By: DinoV

Differential Revision: D52855694

fbshipit-source-id: 57e7034009a567a5e3beacee1f3a1a86b2111ab3