only inline preloaded functions
Summary:
It's not safe to preload a new function mid-compile, since that can run arbitrary Python code and raise a Python exception. So restrict ourselves to only inlining already-preloaded functions.

For batch compile, this is what we were doing already. For single-function compile, try to preload any function loaded as a global that we might call.

Reviewed By: alexmalyshev

Differential Revision: D53369676

fbshipit-source-id: ceea1b2ff660fcfb64b3ee582e23e14007005c3d