fix crasher on dumping JSON passes with inlined function
Summary: When a function is inlined in the HIR inliner, the `BeginInlinedFunction` instruction has access to both the code object of the caller and the callee, and gets the bytecode offset of the call instruction in the caller. We need to make sure it uses the code object of the caller also when getting the line number of the instruction, otherwise it may ask for an out-of-range bytecode offset for the wrong code object, leading to a crash.

Reviewed By: alexmalyshev

Differential Revision: D53557002

fbshipit-source-id: d82a0f8aee5a4adcf199c4babfb0062ee40089b7