defer all imports, error on reference cycle
Summary:
Currently, we sometimes silently fall back to dynamic even when importing a known value from a static module, because of import cycles. This causes unpredictably poorer performance and lack of type checking where we should have it, and is also a problem for dependency tracking.

Fix it by deferring all imports (effectively, make all imports lazy by default) and trigger the declaration visit of a dependency only when we actually reference the value. This means we will only have cyclic problems if we have an actual cyclic module-level reference (which should fail at runtime also.)

Also add an exception when we fail to find an attribute on an incomplete module, since this likely indicates a cycle. This means we will no longer silently fall back to dynamic on cycles.

Reviewed By: pilleye

Differential Revision: D53762347

fbshipit-source-id: 5dbbb2ec6d25b1d7c931c62076dffdad8060aa59