improve inline-import support
Summary:
In working on some other changes, I ran across the fact that a) the `unknown_names` tests are not imported to `__init__.py`, meaning they don't actually run by default (we should really fix our test discovery to not require this), and b) the tests as-is don't provide very good coverage of the related code in `TypeBinder.visitImport` and `TypeBinder.visitImportFrom`.

Given some code like

```
def f():
    from a import b
```

there is code in the import visitors to handle the case where we actually know about the module `a`, and therefore we have type information about `a.b`. But the tests never actually hit this case (and neither do any other tests in our suite), because it relies on something else already having triggered an import of module `a` from the decl visit; there's no code here to actually go look for the `a` module and visit it.

I think that's a little weird; if we're going to support type information about inline imports, it shouldn't depend on someone else already having imported the module. So this diff adds support to actually go visit the module if needed, and updates the tests to ensure we are covering all the cases.

Reviewed By: pilleye

Differential Revision: D53780968

fbshipit-source-id: cf43f0bfc1be62bc4bce449b0a9b92876eb77a7f