simplify ModuleTable.get_child
Summary: The signature of `ModuleTable.get_child` will have to change in the upcoming dependency-tracking work to accept a second "requester" argument which most call-sites will use. Having both this and a `default` arg is confusing at call-sites, and the `default` arg is not really needed; most call sites don't use it, and the few that do can easily be transformed from `get_child("foo", DEFAULT)` to `get_child("foo") or DEFAULT`, with the same semantics.

Reviewed By: pilleye

Differential Revision: D53782083

fbshipit-source-id: 277c13583ccaf732f1808859400b3cb4b2a1272b