Skip zipimport + pdb test
Summary:
We already skip all of test_pdb for JIT testing.  This test slipped by because
it runs pdb in a subprocess.  The CLI arguments in our Makefile are only applied
to our workers, not their subprocesses.  If the JIT is enabled via the PYTHONJIT
env var however, this test will fall over and die.

Reviewed By: oclbdk

Differential Revision: D54542522

fbshipit-source-id: 8c8063ab9ede9b7d0c99eea43b5748c5af0eba50