add support for shadowing in match patterns
Summary:
Adds support for shadowing other variables in match statements. The type of a variable defined in a match statement is dynamic, so if that variable is declared as a different type outside of the match statement, then we cast to that refined type.

This commit works by adding a `type_dict` to `PatternContext` storing a mapping from names defined within a pattern to their Type instance. This type instance is picked up during type binding. Then, when we store, we assume that we're assigning from `dynamic` to that type in the `type_dict`, optionally casting if the destination is not dynamic.

This also fixes a few bugs with visiting that weren't surfaced earlier.

Reviewed By: carljm

Differential Revision: D54497645

fbshipit-source-id: 50c0312689d54abdc890cc5bf1e30f420d20d65b