initial dependency tracking in simple cases
Summary:
Put in place the basics of the dependency-tracking framework, and start actually tracking dependencies in a few simple cases. This is not complete yet, but there's no harm in landing it as-is and working incrementally to make it more complete.

See `cinderx/cinderx/compiler/static/dependency_invalidation.md` for more context.

Reviewed By: DinoV

Differential Revision: D53791981

fbshipit-source-id: 03d42c3b74d44691b495717f24837aca25279ce1