record intra-module dependencies also
Summary:
In previous diffs I was recording only dependencies from one module to another. But we also need to record dependencies within a module, because a cross-module transitive chain of dependencies may rely on having those intra-module dependencies recorded (module A depends on B.f, which depends on B.g, which depends on something from C.) So record them as well.

At this point I'm not aware of gaps in the dependency tracking. Probably will discover some in the future, but for now I'll move on to actually recording the dependencies in pyc files and checking them on load.

Reviewed By: DinoV

Differential Revision: D54612277

fbshipit-source-id: 867bec72016523363308c709ba2fe321d5eb4c7d