find all transitive dependencies of a module
Summary:
Add `ModuleTable.get_dependencies()` method, to get all modules in its transitive dependency closure.

Notably, although the final set of dependencies is just a set of modules, the transitive dependency graph is at function/class granularity. In other words, it is pruned of false dependencies from functions or classes that we don't depend on, which just happen to live in the same module as a function or class we do depend on.

Reviewed By: DinoV

Differential Revision: D54657372

fbshipit-source-id: cabd0cb552e9f8e1259199dc6717ccff66c6b7a9