fix recording deps to imported-as-object unknown modules
Summary: We should be able to record dependencies on unknown modules even when they are imported as an object (e.g. `import unknown_mod`), not only when they are imported-from (e.g. `from unknown_mod import U`). In this case, there's no need to record granular deps, we just need to know we have a dependency on the module.

Reviewed By: DinoV

Differential Revision: D54695279

fbshipit-source-id: 16ba0725ce1031a55af6a269643154ce7505c645