Keep function group until we get to an actual function
Summary:
A pattern can exist where there are a bunch of `overload` functions which are declared and then the function is actually defined via some dynamic behavior.

Currently when this happens we end up hitting a KeyError when we're finishing the bind.

Instead we can always return the `FunctionGroup` and keep that present, and then let that get replaced with the dynamic value.

Reviewed By: vladima

Differential Revision: D54851838

fbshipit-source-id: 4a6ad490a4e4268eae6245f79964944e62dd1feb