Add helper for fetching the name operand of an HIR instruction
Summary:
There's a lot of places in the LIR generator where it loads a string from
co_names via an index, and then immediately constructs an LIR Move to turn it
into an Instruction* value.  Consolidating them into a single helper.

For callsites where it is used as a PyObject* directly (see LoadGlobal) this
will add an extra Move into the instruction stream, but that should get cleaned
up in postalloc.cpp.

Reviewed By: swtaarrs

Differential Revision: D55068784

fbshipit-source-id: 54880a612ec00a0df8e56dff15b6af6d09ead10a