fix Python linting for PythonLib/ move
Summary:
When the CinderX Python code was moved into `cinderx/PythonLib`, we lost linter/formatter coverage on it.

Restore that coverage, and run the linter on all files in PythonLib (`arc lint cinderx/PythonLib/**/*.py`). Looks like a combination of changes in the formatter (I think we recently updated Black) and things that may have snuck through since we lost formatter coverage, as well as adding formatter coverage for the first time to the `__strict__` package (seems it was accidentally missing before.)

Reviewed By: jbower-fb

Differential Revision: D55248992

fbshipit-source-id: 10f1e8684bfb2d362aa4279c600cf06a758d51ee