don't blindly overwrite existing exception in classloader_verify_type
Summary:
We shouldn't overwrite exceptions; it can make debugging harder.

In this case, if we think it's important to add the context "this happened in classloader" we could chain exceptions, but in my experience the underlying exception generally already makes it pretty clear what happened, and it's either not important or already obvious that it's coming from classloader.

I removed a "skip under JIT" decorator from one test. Since my work earlier this half, the JIT can now reliably report errors during preload, so that test works fine under JIT now. But tests that raise errors in JIT preload can't be part of the multithreaded-compile test (since it compiles them outside of the error-catching done in the test), so I made a more targeted fix to exclude them from the multithreaded compile test.

Reviewed By: czardoz

Differential Revision: D55222242

fbshipit-source-id: 8cdf1de18a49ee92e59760bc789a024199cd4a50