add Compiler.get_dependencies()
Summary:
Keep track of details of loaded source files in the compiler, and provide source file details for all dependencies of a static module on request.

Maintain a cache from module-name to SourceInfo, so that dependency validation and actual compilation don't need to stat/read a given source file more than once.

Reviewed By: DinoV

Differential Revision: D55272539

fbshipit-source-id: b9493981483e0dd6eaa5772b0cc40a047f89eb08