Order methods the same way between inline_cache.{cpp,h}
Summary:
The definitions in inline_cache.cpp are all over the place and make it harder to
read.  This is especially true because a lot of the classes here visually look
very similar to each other (e.g. LoadMethodCache vs LoadModuleMethodCache,
LoadAttrCache vs LoadTypeAttrCache).  Order them as they're written out in the
header file.

Found that LoadTypeMethodCache::reset() was declared but never defined, kill it.

Reviewed By: swtaarrs

Differential Revision: D55533991

fbshipit-source-id: b4bfdced4118530b2ae3c38b6fb0ce6df2be7e76