X jit-max-code-size initial implementation
Summary:
Initial go at implementing a limit to the amount of memory the jit compiler will use. The implementation is reactive, it will stop compiling once the limit has been breached; therefore, it will always be breached. This is not reactive, in that it not abort during jit compilation.

This version only tests the cinder allocator and not the slab or none huge page ones. Maybe they are worth testing as a follow on?

Reviewed By: alexmalyshev

Differential Revision: D55872896

fbshipit-source-id: 8e225ef65ec64a3c006e61b09749890e2271c610