Use enum class for ELF section header indices, std::span for tables
Summary:
Makes it easier to track the total number of sections by putting the `kTotal`
value at the end of the enum class.  std::span is a little nicer for tracking a
byte range and it means we don't have to manually compute the size of the symbol
table.

Cleans up the unused variable warning for kNullSectionHeaderIndex as well.

Differential Revision: D55641144

fbshipit-source-id: 03321b14a5cac87f5d0dd1bde51b8c4c59b82ac5