Explicitly link CinderX into C++ tests
Summary:
Breaks the _cinderx extension up into a library with functionality and a Python extension which is just a thin wrapper. This allows us to explicitly statically link CinderX directly into the Runtime/StrictModules tests.

This change is to support Buckifying the C++ tests. The existing setup has `_cinderx.so` as an explicit dynamic library dependency of the test binary, but getting `_cinderx.so` in the right place in a Buck build at execution time is tricky.

Reviewed By: alexmalyshev

Differential Revision: D56214643

fbshipit-source-id: a5e4cfc84596c4bbdcd0dce2dd283af6005c3445