Add awaiter pointer to 3.12
Summary:
There's a good chance we're going to need to forward port the async stack walking implementation
in 3.10 to 3.12. That relies on the presence of the awaiter pointer in coroutines (and we'd like
to support async generators too). Let's get this in early, even if we don't end up needing it,
since it's an ABI breaking change.

Reviewed By: DinoV

Differential Revision: D56263282

fbshipit-source-id: 5539cdb6713beeb8374b473f82e0b227e5f65729