Remove DCHECK from sort comparison function
Summary:
Our implementation of `std::sort()` (and other functions that take a
`Compare` object) does some sanity checks on the provided predicate, including
calling `comp(a, a)` and asserting that it's false. This was tripping a
`JIT_DCHECK()` in this comparison function. Remove it from the predicate and do
a separate pass to ensure uniqueness after sorting.

Reviewed By: alexmalyshev

Differential Revision: D56479803

fbshipit-source-id: e877d08d2f51271fd668fe1308e34aa571b1652b