Explicit alignment for all generated ELF sections
Summary:
Technically a section alignment of 0 is the same as alignment of 1, but most
shared libraries I've inspected use 0 only for the null section.

Right now the only alignment we enforce is 0x1000 for both .text and .dynsym,
but this will change in the future. This diff makes alignUp() more general by
taking the alignment as an argument.

Reviewed By: swtaarrs

Differential Revision: D56472665

fbshipit-source-id: 4a6a94d2a71344dba111324601a149a40db163a6