Better track code buffers for compiled functions
Summary:
After NativeGenerator compiles code, it stores pointers to the function's entry
points and the total size of the code, but it doesn't have a pointer to the
start of the code.  We need this for ELF writing because the goal is to
serialize the entire code buffer, and not just the code that starts at the
normal entry point.

NativeGenerator and CompiledFunction will now expose the code buffer as a span
of (const) bytes.

Differential Revision: D56897617

fbshipit-source-id: 6d6668b7b0b737405c8fff85ff31525ce0614174