Remove most of interpreter type profiling
Summary:
We're not currently using it and it's unclear when we'd be able to invest more
into it.  If we want to come back to feeding better types into the JIT it would
make sense to try to tap into the machinery used by the specializing adaptive
interpreter.

This deletes pretty much everything except ProfileRuntime, which is still used
by the simplifier for LoadAttr on split dicts.  Will get that in a future diff.

Reviewed By: jbower-fb

Differential Revision: D59603172

fbshipit-source-id: 1cd4ca33d4e29faa77e7597e30cd943c65c141a2