Port over gh-91048 for awaiter support in Meta Python 3.12
Summary:
Context: We are deciding to stick with the old "bpf" way of doing async stack walking for Meta Python 3.12, and maybe 3.13. This involves continuing support for the awaiter pointer which is why we need this port. Ideally we get the python stack from the runtime and there is expected to be an internal implementation of this for Meta Python 3.13 or 3.14 latest.

This is essentially a copy-paste port of https://github.com/python/cpython/pull/103976/files from last year. The main change of this diff from the PR is that gi_awaiter is renamed to gi_ci_awaiter

Reviewed By: jbower-fb

Differential Revision: D57072594

fbshipit-source-id: ca0d2ee0ffca2a72e16d479ca395555ed13498af