Redo how CinderX initialization handles getters/setters on callable types
Summary:
The way this works today with 3.10 is that we statically initialize a copy of
the same tables that CPython uses internally.  We had to patch the runtime to
export the getters with `Cix_` prefixes so we could access them in CinderX.

The new way in this diff is to dynamically arrays and then copy the getset
tables into them on CinderX initialization, tacking on the extra
`__typed_signature__` getter.  The arrays are cleaned up by global destructors.

This drops the requirement that we have the `Cix_` wrappers.

The thing that I noticed while working on this is that we never actually reset
these tables when CinderX is unloaded, they stick around until the runtime
process itself dies.  This diff doesn't make that any better, but at least it
tries to document what needs to be modified to make that work.

Reviewed By: jbower-fb

Differential Revision: D59883485

fbshipit-source-id: 204b76465b927a9d78204bcc5346cc9cfb8cc91e