Deliver dict added event only after it's guaranteed to succeed
Summary:
Back port of: https://github.com/python/cpython/pull/122207 fixing https://github.com/python/cpython/issues/122208

The current dictionary watchers implementation delivers the added event before it checks to see if we need to re-size the dictionary. This resize can fail so the value isn't added, and then the tracker is out of sync with the true state of the dictionary.

This moves the delivery of the event to after any necessary allocations have happened.

Reviewed By: jbower-fb

Differential Revision: D60182094

fbshipit-source-id: f34940e98ce1caadeee364f9d126d35839661961