Add space for set-awaiter methods on types without changing existing structs
Summary:
This is part of re-creating the `cr_awaiter` feature backed out in D59866335. The issue was changing the `PyAsyncMethods` struct leads to ABI breakage. In this diff we just make the change to store the extra field we wanted to add in data "outside" of the `PyTypeObject`.

I didn't come up with this; it's mostly a reimplementation for how things were already done in Cinder. We were trying to get away from this while upstreaming this to 3.12 but the patch was rejected.

Reviewed By: alexmalyshev

Differential Revision: D60097673

fbshipit-source-id: 19cabdfec0cfbab35f526494e63a75501b1719d9