Replace _Py_Identifier objects with interned strings
Summary:
_Py_Identifier is internal to CPython and it's been getting harder and harder to
access, see https://bugs.python.org/issue46541.

For most uses in CinderX we can initialize static local variables that will
lazily create interned Python strings.  It's not as efficient as _Py_Identifier,
but it's using the stable API so it'll be easier to maintain over time.

The places where we still have _Py_Identifier relate to using
`_PyObject_LookupSpecial` (in 3.12 it's renamed to `_PyObject_LookupSpecialId`).
There's no equivalent of that function that uses `PyObject*` values, we'll
likely have to implement that ourselves.

Reviewed By: DinoV

Differential Revision: D59936685

fbshipit-source-id: 2393fa5f5fff8205ff5203ce7e814b8e8f340971