Load and initialize CinderX for Python 3.12
Summary:
Loads and initializes the CinderX module as part of Lib/site.py.  If the module
can't be found, such as when CinderX isn't bundled with the installation, then
silently nothing happens.

Failures that bubble up to the top-level CinderX module are no longer
suppressed.  The suppression that does need to be added is to the _static
module, as that's not initializing successfully right now.

I have no idea if this is what we had in mind for initializing CinderX for 3.12.
I believe our plan was to make it more explicit and only load when users added
the extension as a dependency to their TARGETS rules.  But in the meantime this
helps us run tests.

Reviewed By: jbower-fb

Differential Revision: D60384976

fbshipit-source-id: fe28752e967d7562012031540188279d044d66a3