Also check for _cinderx in 3.12/Lib/site.py
Summary:
cinder_binary(
    name = "foo",
    py_version = "3.12",
    ...
  )

will pull in `cinderx`, but not the inner `_cinderx` C++ extension, so it will
fail to initialize.  The `_cinderx` module isn't ready to be built without
an explicit `cinderx.use_3_12=true` buckconfig value, so we can't bundle
it in yet. The easiest way to get out of this hole is to also check for the
existence of the `_cinderx` module.

  python_binary(
    name = "foo",
    py_version = "3.12",
    ...
  )

is fine on the other hand.  It doesn't pull in `cinderx`.

Differential Revision: D60594172

fbshipit-source-id: 4ad6834d5baf720419d95dc99436d1c4a772ce3d