Remove Ci_TypeAttrRaisingCallback
Summary:
Removes the `Ci_TypeAttrRaisingCallback` hook. We use the same dictionary watcher added in D60084797.

Unfortunately we can no longer fail the updates for overriding a final method or assigning a known incompatible type - instead we now just issue a warning.

Unlike the previous version we no longer need to worry about the update failing - we know now that the dictionary watcher event will be delivered only when it can succeed after D60182094.

Differential Revision: D60249517

fbshipit-source-id: 9881b83ce9bd044caf538195732d172b827d616e