JIT support for raise and deopt
Summary:
For deopt this is mostly just updating frame reification to target `_PyInterpreterFrame` rather than `PyFrameObject`. For raising we need to adapt to 3.12 only using a single value to represent an exception rather than: value, type, and traceback.

The deopt tests also execute turning `_PyInterpreterFrame`s into `PyFrameObject`s to make use of the public interpreter API and this seems to work too.

Reviewed By: alexmalyshev

Differential Revision: D61318124

fbshipit-source-id: 11c44f1ec079dd5023eb363a572904d94a59cd01