Don't package the _static shared library for 3.10.cinder or 3.12
Summary:
This stopped being a "proper" Python extension in D59930492.  The module
named "_static" is now initialized by the _cinderx extension.

The build rule still exists though, so it will produce an ELF shared library
that gets packaged up into the 3.10.cinder and 3.12 Python installations.  The
tricky thing is that it can't be imported like other Python modules, it doesn't
have the prerequisite symbols.  Trying to import it results in an opaque
"invalid ELF header" error.

Right now this works in 3.10.cinder because _cinderx will initialize the _static
module first before it ever gets imported.  So once the first `import _static`
statement is reached, it just gets fetched from `sys.modules`.  However in 3.12
the module currently fails to initialize (T194028831) but the _cinderx
initialization silently continues on.  Once the first `import _static` is
reached, the runtime tries to load the packaged shared library and dies with the
non-obvious error.

Reviewed By: mpage

Differential Revision: D62300617

fbshipit-source-id: 6b1b7fc6a975d584c145e047c111a1668e77e372